WIP migrating to new gpusync.h
gpusync.h had to be updated in order to support threading. This change
updates StereoDisparity (both C and ZC) to use the new format. The
remaining benchmarks still need to be updated to fit the interface.
Next, benchmark.c needs to be updated to invoke it properly. Finally,
the remaining benchmarks can be updated to be threadable.